Move kueuectl-docs to cmd
