+++ cmd/kueuectl-docs/templates/main.html
@@ -7,7 +7,7 @@ no_list: {{.NoList}}
 
 {{ html `<!--
 The file is auto-generated from the Go source code of the component using the
-[generator](https://github.com/kubernetes-sigs/kueue/tree/main/hack/internal/tools/kueuectl-docs).
+[generator](https://github.com/kubernetes-sigs/kueue/tree/cmd/kueuectl-docs).
 -->`}}
 
 ## Synopsis
